refactor: remove django relevant codes
