+++ solvers/utilsx/__init__.py
@@ -4,8 +4,7 @@ import importlib
 from types import ModuleType
 from typing import Dict, List
 
-from static.consts import types as PUZZLE_TYPES
-
+from .consts import types as PUZZLE_TYPES
 from .encoding import Encoding, decode, encode
 
 modules: Dict[str, ModuleType] = {}
