+++ solvers/utilsx/consts.py
